requirements.txt
fastapi
aioredis
uvicorn
tortoise-orm[asyncpg]
python-dotenv
redis
jupyterlab

diffusers
transformers
accelerate
safetensors

torch
torchvision
matplotlib
requests
bitsandbytes
